Improve the device detection (at least on Mac, haven't tested on other platforms)
